requirements.txt
python-telegram-bot[job-queue]>=20.7
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.29
httpx>=0.27.0
python-dotenv>=1.0.1
pytest>=8.2.0
pytest-asyncio>=0.23.6
playwright>=1.44.0
pycookiecheat>=0.8.0
